Bug fixed
After attempting more than 9 attempts the game was still asking for guess instead of printing "GAME OVER" to console.
It was happening because you are checking if a < n first and then checking if count == 9. So, if first condition gets true, it will not check for second condition and the program will keep running even if the count is equal to 9.